# docs/swagger.yaml
definitions:
  models.Song:
    properties:
      group_name:
        type: string
      id:
        type: string
      link:
        type: string
      release_date:
        type: string
      song_name:
        type: string
      text:
        type: string
    required:
    - group_name
    - song_name
    type: object
info:
  contact: {}
paths:
  /songs:
    get:
      description: Get all songs
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Song'
            type: array
      summary: Get list of songs
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Create song by its ID
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Create song by ID
      tags:
      - songs
  /songs/{id}:
    get:
      consumes:
      - application/json
      description: Get song by its ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Get song by ID
      tags:
      - songs
    post:
      consumes:
      - application/json
      description: Update song by its ID
      parameters:
      - description: Song ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Song'
        "404":
          description: Not Found
          schema:
            type: string
      summary: Put song by ID
      tags:
      - songs
swagger: "2.0"